refactor: Product 클래스의 BaseEntity 상속 삭제, quantity 자료형 변경

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/CreateProductRequest.java
@@ -16,7 +16,7 @@ public record CreateProductRequest(
         Long price,
 
         @NotNull
-        Long quantity,
+        int quantity,
 
         @NotNull
         ProductStatus status,
