+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/CreateProductResponse.java
@@ -10,7 +10,7 @@ public record CreateProductResponse(
         String name,
         String category,
         Long price,
-        Long quantity,
+        int quantity,
         ProductStatus status,
         LocalDateTime createdAt,
         Long managerId
